Fix 32 bit shifting issue

+++ spi1_alt2/serial.c
@@ -180,7 +180,7 @@ uint16_t serial_rx_word() {
 }
 
 uint32_t serial_rx_long() {
-  return serial_rx_word() << 16 + serial_rx_word();
+  return ((uint32_t)serial_rx_word() << 16) + serial_rx_word();
 }
 
 void serial_tx_byte(uint8_t tx_byte) {
