+++ uart1_alt2/serial.c
@@ -44,8 +44,9 @@ uint16_t serial_rx_word() {
 }
 
 uint32_t serial_rx_long() {
-  return serial_rx_word() << 16 + serial_rx_word();
+  return ((uint32_t)serial_rx_word() << 16) + serial_rx_word();
 }
+
 
 void serial_tx_byte(uint8_t tx_byte) {
   UTX1IF = 0;
